requirements.txt
numpy
matplotlib
scikit-learn
cupy-cuda12x
